fix: playlist naddr parsing and consolidate duplicated parse_naddr functions
The playlist detail page split the naddr parameter on colons, but callers
pass bech32-encoded naddr strings (naddr1...) which contain no colons,
causing an immediate 'Invalid playlist address' error. Fix by using proper
bech32 decoding via a shared parse_naddr utility in utils/nip19.rs.

Consolidate 6 duplicated parse_naddr/parse_coordinate/decode_naddr
functions across track_detail, live_stream_detail, podcast_episode_detail,
podcast_nostr_detail, and article_detail into the shared utility that
handles both bech32 naddr and colon-separated formats.

Closes #313
