+++ src/routes/podcast/podcast_nostr_detail.rs
@@ -384,7 +384,7 @@ fn PodcastDetailSkeleton() -> Element {
 async fn fetch_nostr_podcast_metadata(
     naddr: &str,
 ) -> std::result::Result<(PodcastMetadata, String), String> {
-    let (pubkey, d_tag) = parse_coordinate(naddr)?;
+    let (pubkey, d_tag) = crate::utils::nip19::parse_naddr(naddr)?;
     let metadata_filter = Filter::new()
         .kind(Kind::from(podcast::KIND_APP_DATA))
         .author(PublicKey::from_hex(&pubkey).map_err(|e| e.to_string())?)
@@ -433,27 +433,4 @@ async fn fetch_nostr_episodes(
     Ok(episodes)
 }
 
-/// Parse coordinate string into pubkey and d-tag
-fn parse_coordinate(coord: &str) -> Result<(String, String), String> {
-    use nostr::prelude::*;
-    if coord.starts_with("naddr") {
-        let nip19 = Nip19::from_bech32(coord).map_err(|e| format!("Invalid naddr: {}", e))?;
-        match nip19 {
-            Nip19::Coordinate(nip19_coord) => {
-                let pubkey = nip19_coord.coordinate.public_key.to_hex();
-                let d_tag = nip19_coord.coordinate.identifier;
-                Ok((pubkey, d_tag))
-            }
-            _ => Err("Expected naddr coordinate".to_string()),
-        }
-    } else {
-        let parts: Vec<&str> = coord.splitn(3, ':').collect();
-        if parts.len() >= 3 {
-            return Ok((parts[1].to_string(), parts[2].to_string()));
-        }
-        if parts.len() == 2 {
-            return Ok((parts[0].to_string(), parts[1].to_string()));
-        }
-        Err(format!("Invalid coordinate format: {}", coord))
-    }
-}
+
