+++ src/routes/video/live_stream_detail.rs
@@ -6,11 +6,13 @@ use crate::stores::nostr_client::{fetch_events_aggregated, CLIENT_INITIALIZED, H
 use crate::stores::profiles;
 use crate::utils::truncate_pubkey;
 use dioxus::prelude::*;
-use nostr_sdk::{Filter, FromBech32, Kind, PublicKey};
+use nostr_sdk::{Filter, Kind, PublicKey};
 use std::time::Duration;
 #[component]
 pub fn LiveStreamDetail(note_id: String) -> Element {
-    let parsed_naddr = use_memo(move || parse_naddr(&note_id));
+    let parsed_naddr = use_memo(move || {
+        crate::utils::nip19::parse_naddr(&note_id).unwrap_or((String::new(), String::new()))
+    });
     let mut stream_event = use_signal(|| None::<nostr_sdk::Event>);
     let mut stream_meta = use_signal(|| None::<LiveStreamMeta>);
     let mut loading = use_signal(|| true);
@@ -356,25 +358,6 @@ pub fn LiveStreamDetail(note_id: String) -> Element {
                     }
                 }
             }
-        }
-    }
-}
-/// Parse naddr format - supports both NIP-19 bech32 and "30311:pubkey:dtag" formats
-fn parse_naddr(note_id: &str) -> (String, String) {
-    if let Ok(nostr_sdk::nips::nip19::Nip19::Coordinate(coord)) =
-        nostr_sdk::nips::nip19::Nip19::from_bech32(note_id)
-    {
-        return (coord.public_key.to_hex(), coord.identifier.clone());
-    }
-    let parts: Vec<&str> = note_id.splitn(3, ':').collect();
-    if parts.len() >= 3 {
-        (parts[1].to_string(), parts[2].to_string())
-    } else {
-        let parts: Vec<&str> = note_id.splitn(2, ':').collect();
-        if parts.len() == 2 {
-            (parts[0].to_string(), parts[1].to_string())
-        } else {
-            (String::new(), String::new())
         }
     }
 }
